+++ include/FileClasses/xmidi/databuf.h
@@ -401,13 +401,13 @@ public:
 
     virtual unsigned int getSize() {
         unsigned int oldPos = getPos();
-        unsigned int size = SDL_RWseek(rwop,0,SEEK_END);
+        unsigned int size = static_cast<unsigned int>(SDL_RWseek(rwop,0,SEEK_END));
         SDL_RWseek(rwop,oldPos,SEEK_SET);
         return size;
     };
 
     virtual unsigned int getPos() {
-        return SDL_RWtell(rwop);
+        return static_cast<unsigned int>(SDL_RWtell(rwop));
     };
 };
